omics_ml_pipeline/requirements.txt
pandas>=2.0
numpy>=1.26
scikit-learn>=1.4
xgboost>=2.0
mlflow>=2.13
hyperopt>=0.2.7
pyyaml>=6.0
matplotlib>=3.8
scipy>=1.12
rich>=13.0
